refactor: move slugger to metadata generator dir
